Major Updates
added settings file to collect global variables.
improved logging with times and more accurate error reports.
fixed broken tools.
changed how docker works.

+++ BotHead.py
@@ -1,81 +1,59 @@
-import discord
 from discord.ext import commands
+import settings
+import discord
+import datetime
 import os
-from tools.basicTools import readJson
-import sys
-import mysql.connector
 
-testing = True
-
-if len(sys.argv) == 2:
-    if "false" in sys.argv[1] or "False" in sys.argv[1]:
-        testing = False
-
-db = mysql.connector.connect(
-    host="192.168.1.250",
-    user="discord",
-    password="plop9100",
-    database="discord"
-)
-
-my_cursor = db.cursor()
-token = None
-
-if testing:
-    sql = "SELECT * FROM tokens WHERE name ='Test'"
-    my_cursor.execute(sql)
-    my_result = my_cursor.fetchall()
-    for x in my_result:
-        token = x[1]
-else:
-    sql = "SELECT * FROM tokens WHERE name ='Live'"
-    my_cursor.execute(sql)
-    my_result = my_cursor.fetchall()
-    for x in my_result:
-        token = x[1]
-
-json_name = "info.json"
-json = readJson(json_name)
 client = commands.Bot(command_prefix='.')
-
-for f in os.listdir('./cogs'):
-    if f.endswith('.py'):
-        client.load_extension(f'cogs.{f[:-3]}')
 
 
 @client.command(brief="Admin only command: Load a Cog.")
 async def load(ctx, extension):
-    client.add_cog(f'cogs.{extension}')
-    print(f'cog:{extension} loaded')
-    await ctx.message.delete()
+    print(f"""{datetime.datetime.now()}: load from {ctx.author} """)
+    if str(ctx.message.channel) in settings.json["command_channels"]:
+        client.add_cog(f'{extension}')
+        print(f'cog:{extension} loaded')
+        await ctx.message.delete()
 
 
 @client.command(brief="Admin only command: Unload a Cog.")
 async def unload(ctx, extension):
-    client.remove_cog(f'cogs.{extension}')
-    print(f'cog:{extension} unloaded')
+    print(f"""{datetime.datetime.now()}: unload from {ctx.author}""")
+    if str(ctx.message.channel) in settings.json["command_channels"]:
+        client.remove_cog(f'{extension}')
+        print(f'cog:{extension} unloaded')
     await ctx.message.delete()
 
 
 @client.command(brief="Admin only command: Reload a Cog.")
 async def reload(ctx, extension):
-    client.remove_cog(f'cogs.{extension}')
-    print(f'cog:{extension} unloaded')
-    client.add_cog(f'cogs.{extension}')
-    print(f'cog:{extension} loaded')
+    print(f"""{datetime.datetime.now()}: reload from {ctx.author}""")
+    if str(ctx.message.channel) in settings.json["command_channels"]:
+        client.remove_cog(f'{extension}')
+        print(f'cog:{extension} unloaded')
+        client.add_cog(f'{extension}')
+        print(f'cog:{extension} loaded')
     await ctx.message.delete()
 
 
 @client.event
 async def on_ready():
-    print(f"""Logged on as {client.user}!""")
+    print(f"""{datetime.datetime.now()}: Logged on as {client.user}!""")
     await client.change_presence(status=discord.Status.online, activity=discord.Game('listening'))
 
 
 @client.event
 async def on_disconnect():
+    print(f"""{datetime.datetime.now()}: Logged off!""")
     await client.change_presence(status=discord.Status.offline, activity=discord.Game('sleeping'))
 
 
-if token:
-    client.run(token)
+if __name__ == "__main__":
+
+    settings.init(db_host="192.168.1.250", db_username="discord", db_password="plop9100")
+
+    for f in os.listdir('./cogs'):
+        if f.endswith('.py'):
+            client.load_extension(f'cogs.{f[:-3]}')
+
+    client.run(settings.token)
